refactor(search): intermediate step of renaming components
Intermediate step of renaming components
