+++ src/components/UI/icons/RemoveSearch111.js
